+++ quick_blue_platform_interface/lib/quick_blue_platform_interface.dart
@@ -5,3 +5,4 @@ export 'src/bluetooth_gatt.dart';
 export 'src/callbacks.dart';
 export 'src/quick_blue_exception.dart';
 export 'src/quick_blue_platform.dart';
+export 'src/scan_filter.dart';